Made review changes

+++ core/controllers/question_editor.py
@@ -115,31 +115,26 @@ class QuestionSkillLinkHandler(base.BaseHandler):
     """A handler for linking and unlinking questions to or from a skill."""
 
     @acl_decorators.can_manage_question_skill_status
-    def put(self, question_id, skill_ids):
-        """Updates the difficulty of the question with respect to the given
-        skill.
+    def put(self, question_id):
+        """Updates the QuestionSkillLink models with respect to the given
+        question.
         """
-        skill_ids = skill_ids.split(',')
-
         if self.payload.get('action') == 'update_difficulty':
             new_difficulty = float(self.payload.get('new_difficulty'))
+            skill_id = self.payload.get('skill_id')
             question_services.update_question_skill_link_difficulty(
-                question_id, skill_ids[0], new_difficulty)
-        elif self.payload.get('action') == 'add_skill':
+                question_id, skill_id, new_difficulty)
+        elif self.payload.get('action') == 'edit_links':
             difficulty = self.payload.get('difficulty')
-            for skill_id in skill_ids:
-                question_services.create_new_question_skill_link(
-                    self.user_id, question_id, skill_id, difficulty)
+            skill_ids_task_list = self.payload.get('skill_ids_task_list')
+            for task_dict in skill_ids_task_list:
+                if task_dict['task'] == 'remove':
+                    question_services.delete_question_skill_link(
+                        self.user_id, question_id, task_dict['id'])
+                elif task_dict['task'] == 'add':
+                    question_services.create_new_question_skill_link(
+                        self.user_id, question_id, task_dict['id'], difficulty)
 
-        self.render_json(self.values)
-
-    @acl_decorators.can_manage_question_skill_status
-    def delete(self, question_id, skill_ids):
-        """Unlinks a question from a skill."""
-        skill_ids = skill_ids.split(',')
-        for skill_id in skill_ids:
-            question_services.delete_question_skill_link(
-                self.user_id, question_id, skill_id)
         self.render_json(self.values)
